feat: 검증된 Player 모델 기능 구현
