+++ README.md
@@ -40,7 +40,7 @@
 ---
 - 플레이어
   - 플레이어는 이름이랑 블랙잭을 가지고 있다.
-    - [ ] 이름은 중복 될 수 없다.
+    - [x] 이름은 중복 될 수 없다.
 - 딜러
   - 딜러는 이름과 블랙잭을 가지고 있다.
     - [ ] 딜러의 이름은 생성자로 구현된다. (딜러)
